add user checks

+++ assets/router.py
@@ -1,10 +1,11 @@
 from typing import List
 
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends
 from starlette.background import BackgroundTasks
 
 from assets.db import collection as asset_collection
 from assets.models import Asset
+from auth.dependencies import get_current_active_user
 from common.utils import aggregate_transactions
 from transactions.db import collection as transaction_collection
 
@@ -12,13 +13,16 @@ router = APIRouter(prefix="/assets")
 
 
 @router.get("/", response_model=List[Asset])
-async def get_assets(background_tasks: BackgroundTasks):
-    result = asset_collection.find({})
+async def get_assets(
+    background_tasks: BackgroundTasks, user=Depends(get_current_active_user)
+):
+    query = {"userId": user.uid}
+    result = asset_collection.find(query)
     assets = []
     for doc in result:
         assets.append(Asset.parse_obj(doc))
 
-    background_tasks.add_task(aggregate_transactions)
+    background_tasks.add_task(aggregate_transactions, query)
     return assets
 
 
@@ -34,7 +38,7 @@ async def update_asset(asset_id: str):
 
 
 @router.delete("/{asset_id}/")
-async def delete_asset(asset_id: str):
-    transaction_collection.delete_many({"assetId": asset_id})
-    asset_collection.delete_one({"id": asset_id})
+async def delete_asset(asset_id: str, user=Depends(get_current_active_user)):
+    transaction_collection.delete_many({"assetId": asset_id, "userId": user.uid})
+    asset_collection.delete_one({"id": asset_id, "userId": user.uid})
     return ""
